+++ srcs/mandatory/headers/fractal.h
@@ -6,7 +6,7 @@
 /*   By: seojilee <seojilee@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/02 19:42:45 by seojilee          #+#    #+#             */
-/*   Updated: 2023/12/14 08:49:17 by seojilee         ###   ########.fr       */
+/*   Updated: 2023/12/14 15:17:23 by seojilee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,14 +28,6 @@
 # define ITER_SMALL		100
 # define ITER			500
 # define ITER_LARGE		1000
-
-typedef struct s_matrix {
-	double	mat[3][3];
-}		t_matrix;
-
-typedef struct s_vec3d {
-	double	vec[3];
-}		t_vec3d;
 
 typedef struct s_julia3d {
 	int	x;
@@ -64,6 +56,9 @@ void		draw_julia(double z_abs, t_data *img, int i, int j);
 void		draw_burningship(double z_abs, t_data *img, int i, int j);
 void		draw_tricorn(double z_abs, t_data *img, int i, int j);
 
+// fractal3d_recurrence.c
+void		iter_julia3d(t_julia3d *dots, t_complex c, int z_value);
+
 // logistic_utils
 bool		check_inbox(t_xy xy);
 void		make_box_black(t_data *img);
@@ -72,16 +67,10 @@ void		logistic_next(double r, double *x);
 // julia3d_utils
 bool		all_diverge(t_complex c);
 int			get_total_slices(t_complex c);
-void		iter_julia3d(t_julia3d *dots, t_complex c, int z_value);
-void		init_matrix(t_matrix *matrix);
-void		rotation_per_axis(t_matrix *matx, t_matrix *maty, \
-		t_matrix *matz, t_vec3d angle);
-t_matrix	matrix_multiply(t_matrix mat1, t_matrix mat2);
-t_matrix	generate_rotation_matrix(t_data *img);
-void		matrix_point_multiply(t_matrix mat, t_julia3d *dots, int u);
-void		draw_julia3d(t_data *img, t_julia3d *dots, \
-		int total_dots);
+void		set_dot_values(t_julia3d *dot, int x_value, int y_value, int z_value);
 int			find_layer(t_data *img, t_julia3d *dots);
+void		draw_julia3d(t_data *img, t_julia3d *dots, \
+				int total_dots);
 
 void		mandelbrot(t_data *img);
